# api/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/v1/me": {
      "get": {
        "operationId": "MeController_me",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Profil et droits de l’utilisateur connecté",
        "tags": [
          "Session"
        ]
      }
    },
    "/v1/me/navigation": {
      "get": {
        "operationId": "MeController_navigation",
        "parameters": [
          {
            "name": "accept-language",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Menu dynamique : entrées autorisées, dans la langue demandée",
        "tags": [
          "Session"
        ]
      }
    },
    "/v1/me/mfa/setup": {
      "post": {
        "operationId": "MeController_setupMfa",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Génère le secret TOTP à scanner dans l’application d’authentification",
        "tags": [
          "Session"
        ]
      }
    },
    "/v1/me/mfa/enable": {
      "post": {
        "operationId": "MeController_enableMfa",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MfaCodeDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Active la MFA après vérification d’un premier code",
        "tags": [
          "Session"
        ]
      }
    },
    "/v1/navigation-items": {
      "get": {
        "operationId": "NavigationController_list",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Toutes les entrées du menu (paramétrage)",
        "tags": [
          "Administration"
        ]
      },
      "post": {
        "operationId": "NavigationController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateNavigationItemDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Administration"
        ]
      }
    },
    "/v1/navigation-items/{id}": {
      "get": {
        "operationId": "NavigationController_get",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Administration"
        ]
      },
      "patch": {
        "operationId": "NavigationController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateNavigationItemDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Modifier une entrée (If-Match obligatoire)",
        "tags": [
          "Administration"
        ]
      },
      "delete": {
        "operationId": "NavigationController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Administration"
        ]
      }
    },
...
    "/v1/guests": {
      "get": {
        "operationId": "GuestsController_list",
        "parameters": [
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 200,
              "default": 50,
              "type": "number"
            }
          },
          {
            "name": "cursor",
            "required": false,
            "in": "query",
            "description": "Jeton `nextCursor` de la page précédente",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "fields",
            "required": false,
            "in": "query",
            "description": "Champs à renvoyer, séparés par des virgules (ex. id,status,guest)",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "q",
            "required": false,
            "in": "query",
            "description": "Recherche sur nom, prénom ou email",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "segment",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "individual",
                "group",
                "company"
              ]
            }
          },
          {
            "name": "sort",
            "required": false,
            "in": "query",
            "description": "last_name | -last_name | created_at | -created_at",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Clients"
        ]
      },
      "post": {
        "operationId": "GuestsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateGuestDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Clients"
        ]
      }
    },
    "/v1/guests/{id}": {
      "get": {
        "operationId": "GuestsController_get",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Clients"
        ]
      },
      "patch": {
        "operationId": "GuestsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "if-match",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "If-Match",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateGuestDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Clients"
        ]
      }
    },
    "/v1/guests/{id}/stays": {
      "get": {
        "operationId": "GuestsController_stays",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 200,
              "default": 50,
              "type": "number"
            }
          },
          {
            "name": "cursor",
            "required": false,
            "in": "query",
            "description": "Jeton `nextCursor` de la page précédente",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "fields",
            "required": false,
            "in": "query",
            "description": "Champs à renvoyer, séparés par des virgules (ex. id,status,guest)",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Historique des séjours du client",
        "tags": [
          "Clients"
        ]
      }
    },
    "/v1/guests/{id}/erase": {
      "post": {
        "operationId": "GuestsController_erase",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Effacement RGPD (anonymisation irréversible)",
        "tags": [
          "Clients"
        ]
      }
    },
...
  },
  "info": {
    "title": "API – Système de gestion hôtelière",
    "description": "Socle V1 : session, menu dynamique, chambres, clients, disponibilités, réservations, réception.",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "MfaCodeDto": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "example": "123456"
          }
        },
        "required": [
          "code"
        ]
      },
      "CreateNavigationItemDto": {
        "type": "object",
        "properties": {
          "key": {
            "type": "string",
            "example": "reservations"
          },
          "parentKey": {
            "type": "object",
            "example": "frontdesk"
          },
          "path": {
            "type": "object",
            "example": "/reservations"
          },
          "icon": {
            "type": "object",
            "example": "calendar"
          },
          "permission": {
            "type": "string",
            "enum": [
              "reservations:read",
              "reservations:write",
              "frontdesk:operate",
              "availability:read",
              "guests:read",
              "guests:write",
              "guests:erase",
              "rooms:read",
              "rooms:write",
              "rooms:status",
              "reports:read",
              "admin:navigation",
              "admin:users",
              "billing:read",
              "billing:write",
              "billing:charge",
              "billing:override",
              "billing:export",
              "settings:billing"
            ]
          },
          "module": {
            "type": "string",
            "example": "core",
            "default": "core"
          },
          "labelFr": {
            "type": "string"
          },
          "labelEn": {
            "type": "string"
          },
          "labelAr": {
            "type": "string"
          },
          "sortOrder": {
            "type": "number",
            "default": 0
          },
          "enabled": {
            "type": "boolean",
            "default": true
          }
        },
        "required": [
          "key",
          "labelFr",
          "labelEn",
          "labelAr"
        ]
      },
      "UpdateNavigationItemDto": {
        "type": "object",
        "properties": {
          "key": {
            "type": "string",
            "example": "reservations"
          },
          "parentKey": {
            "type": "object",
            "example": "frontdesk"
          },
          "path": {
            "type": "object",
            "example": "/reservations"
          },
          "icon": {
            "type": "object",
            "example": "calendar"
          },
          "permission": {
            "type": "string",
            "enum": [
              "reservations:read",
              "reservations:write",
              "frontdesk:operate",
              "availability:read",
              "guests:read",
              "guests:write",
              "guests:erase",
              "rooms:read",
              "rooms:write",
              "rooms:status",
              "reports:read",
              "admin:navigation",
              "admin:users",
              "billing:read",
              "billing:write",
              "billing:charge",
              "billing:override",
              "billing:export",
              "settings:billing"
            ]
          },
          "module": {
            "type": "string",
            "example": "core",
            "default": "core"
          },
          "labelFr": {
            "type": "string"
          },
          "labelEn": {
            "type": "string"
          },
          "labelAr": {
            "type": "string"
          },
          "sortOrder": {
            "type": "number",
            "default": 0
          },
          "enabled": {
            "type": "boolean",
            "default": true
          }
        }
      },
...
      "CreateGuestDto": {
        "type": "object",
        "properties": {
          "firstName": {
            "type": "string",
            "example": "Amina"
          },
          "lastName": {
            "type": "string",
            "example": "Benali"
          },
          "email": {
            "type": "string"
          },
          "phone": {
            "type": "string",
            "example": "+33612345678"
          },
          "nationality": {
            "type": "string",
            "example": "FR",
            "description": "Code pays ISO 3166-1 alpha-2"
          },
          "segment": {
            "type": "string",
            "enum": [
              "individual",
              "group",
              "company"
            ]
          },
          "preferences": {
            "type": "object",
            "example": {
              "pillow": "firm",
              "floor": "high"
            }
          }
        },
        "required": [
          "firstName",
          "lastName"
        ]
      },
      "UpdateGuestDto": {
        "type": "object",
        "properties": {
          "firstName": {
            "type": "string",
            "example": "Amina"
          },
          "lastName": {
            "type": "string",
            "example": "Benali"
          },
          "email": {
            "type": "string"
          },
          "phone": {
            "type": "string",
            "example": "+33612345678"
          },
          "nationality": {
            "type": "string",
            "example": "FR",
            "description": "Code pays ISO 3166-1 alpha-2"
          },
          "segment": {
            "type": "string",
            "enum": [
              "individual",
              "group",
              "company"
            ]
          },
          "preferences": {
            "type": "object",
            "example": {
              "pillow": "firm",
              "floor": "high"
            }
          }
        }
      },
...
    }
  }
}